remove androcov module from droidbot and add a CoverageEvaluator
